+++ core/src/main/java/bisq/core/grpc/CoreApi.java
@@ -17,8 +17,6 @@
 
 package bisq.core.grpc;
 
-import bisq.core.btc.Balances;
-import bisq.core.btc.wallet.WalletsManager;
 import bisq.core.monetary.Price;
 import bisq.core.offer.CreateOfferService;
 import bisq.core.offer.Offer;
@@ -32,27 +30,16 @@ import bisq.core.trade.statistics.TradeStatisticsManager;
 import bisq.core.user.User;
 
 import bisq.common.app.Version;
-import bisq.common.util.Tuple2;
 
 import org.bitcoinj.core.Coin;
-import org.bitcoinj.crypto.KeyCrypterScrypt;
 
 import javax.inject.Inject;
-
-import org.spongycastle.crypto.params.KeyParameter;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import lombok.extern.slf4j.Slf4j;
-
-import javax.annotation.Nullable;
-
-import static bisq.core.grpc.ApiStatus.*;
-import static java.util.concurrent.TimeUnit.SECONDS;
 
 /**
  * Provides high level interface to functionality of core Bisq features.
@@ -60,54 +47,27 @@ import static java.util.concurrent.TimeUnit.SECONDS;
  */
 @Slf4j
 public class CoreApi {
-    private final Balances balances;
     private final OfferBookService offerBookService;
     private final TradeStatisticsManager tradeStatisticsManager;
     private final CreateOfferService createOfferService;
     private final OpenOfferManager openOfferManager;
-    private final WalletsManager walletsManager;
     private final User user;
 
-    @Nullable
-    private String tempLockWalletPassword;
-
     @Inject
-    public CoreApi(Balances balances,
-                   OfferBookService offerBookService,
+    public CoreApi(OfferBookService offerBookService,
                    TradeStatisticsManager tradeStatisticsManager,
                    CreateOfferService createOfferService,
                    OpenOfferManager openOfferManager,
-                   WalletsManager walletsManager,
                    User user) {
-        this.balances = balances;
         this.offerBookService = offerBookService;
         this.tradeStatisticsManager = tradeStatisticsManager;
         this.createOfferService = createOfferService;
         this.openOfferManager = openOfferManager;
-        this.walletsManager = walletsManager;
         this.user = user;
     }
 
     public String getVersion() {
         return Version.VERSION;
-    }
-
-    public Tuple2<Long, ApiStatus> getAvailableBalance() {
-        if (!walletsManager.areWalletsAvailable())
-            return new Tuple2<>(-1L, WALLET_NOT_AVAILABLE);
-
-        if (walletsManager.areWalletsEncrypted())
-            return new Tuple2<>(-1L, WALLET_IS_ENCRYPTED_WITH_UNLOCK_INSTRUCTION);
-
-        try {
-            long balance = balances.getAvailableBalance().get().getValue();
-            return new Tuple2<>(balance, OK);
-        } catch (Throwable t) {
-            // TODO Derive new ApiStatus codes from server stack traces.
-            t.printStackTrace();
-            // TODO Fix bug causing NPE thrown by getAvailableBalance().
-            return new Tuple2<>(-1L, INTERNAL);
-        }
     }
 
     public List<TradeStatistics2> getTradeStatistics() {
@@ -185,96 +145,4 @@ public class CoreApi {
                 log::error);
     }
 
-    // Provided for automated wallet protection method testing, despite the
-    // security risks exposed by providing users the ability to decrypt their wallets.
-    public Tuple2<Boolean, ApiStatus> removeWalletPassword(String password) {
-        if (!walletsManager.areWalletsAvailable())
-            return new Tuple2<>(false, WALLET_NOT_AVAILABLE);
-
-        if (!walletsManager.areWalletsEncrypted())
-            return new Tuple2<>(false, WALLET_NOT_ENCRYPTED);
-
-        KeyCrypterScrypt keyCrypterScrypt = walletsManager.getKeyCrypterScrypt();
-        if (keyCrypterScrypt == null)
-            return new Tuple2<>(false, WALLET_ENCRYPTER_NOT_AVAILABLE);
-
-        KeyParameter aesKey = keyCrypterScrypt.deriveKey(password);
-        if (!walletsManager.checkAESKey(aesKey))
-            return new Tuple2<>(false, INCORRECT_WALLET_PASSWORD);
-
-        walletsManager.decryptWallets(aesKey);
-        return new Tuple2<>(true, OK);
-    }
-
-    public Tuple2<Boolean, ApiStatus> setWalletPassword(String password, String newPassword) {
-        try {
-            if (!walletsManager.areWalletsAvailable())
-                return new Tuple2<>(false, WALLET_NOT_AVAILABLE);
-
-            KeyCrypterScrypt keyCrypterScrypt = walletsManager.getKeyCrypterScrypt();
-            if (keyCrypterScrypt == null)
-                return new Tuple2<>(false, WALLET_ENCRYPTER_NOT_AVAILABLE);
-
-            if (newPassword != null && !newPassword.isEmpty()) {
-                // TODO Validate new password before replacing old password.
-                if (!walletsManager.areWalletsEncrypted())
-                    return new Tuple2<>(false, WALLET_NOT_ENCRYPTED);
-
-                KeyParameter aesKey = keyCrypterScrypt.deriveKey(password);
-                if (!walletsManager.checkAESKey(aesKey))
-                    return new Tuple2<>(false, INCORRECT_OLD_WALLET_PASSWORD);
-
-                walletsManager.decryptWallets(aesKey);
-                aesKey = keyCrypterScrypt.deriveKey(newPassword);
-                walletsManager.encryptWallets(keyCrypterScrypt, aesKey);
-                return new Tuple2<>(true, OK);
-            }
-
-            if (walletsManager.areWalletsEncrypted())
-                return new Tuple2<>(false, WALLET_IS_ENCRYPTED);
-
-            // TODO Validate new password.
-            KeyParameter aesKey = keyCrypterScrypt.deriveKey(password);
-            walletsManager.encryptWallets(keyCrypterScrypt, aesKey);
-            return new Tuple2<>(true, OK);
-        } catch (Throwable t) {
-            // TODO Derive new ApiStatus codes from server stack traces.
-            t.printStackTrace();
-            return new Tuple2<>(false, INTERNAL);
-        }
-    }
-
-    public Tuple2<Boolean, ApiStatus> lockWallet() {
-        if (tempLockWalletPassword != null) {
-            Tuple2<Boolean, ApiStatus> encrypted = setWalletPassword(tempLockWalletPassword, null);
-            tempLockWalletPassword = null;
-            if (!encrypted.second.equals(OK))
-                return encrypted;
-
-            return new Tuple2<>(true, OK);
-        }
-        return new Tuple2<>(false, WALLET_ALREADY_LOCKED);
-    }
-
-    public Tuple2<Boolean, ApiStatus> unlockWallet(String password, long timeout) {
-        Tuple2<Boolean, ApiStatus> decrypted = removeWalletPassword(password);
-        if (!decrypted.second.equals(OK))
-            return decrypted;
-
-        TimerTask timerTask = new TimerTask() {
-            @Override
-            public void run() {
-                log.info("Locking wallet");
-                setWalletPassword(password, null);
-                tempLockWalletPassword = null;
-            }
-        };
-        Timer timer = new Timer("Lock Wallet Timer");
-        timer.schedule(timerTask, SECONDS.toMillis(timeout));
-
-        // Cache wallet password for timeout (secs), in case
-        // user wants to lock the wallet for timeout expires.
-        tempLockWalletPassword = password;
-        return new Tuple2<>(true, OK);
-    }
 }
